openjdk-7-release-03b147.inc: Ignore deprecation
readdir_r has been deprecated since glibc 2.24. As openjdk uses this
function, a warning is generated during build that gets turned into
an error when -Werror is used.

While the readdir function is suggested as an replacement, it has
slightly different API. The warning is ignored instead to prevent any
unexpected breakage from patching the offending functions.

(Note: This is just a port of 0f127a49696 from Erkka Kääriä for openjdk-7
, all the credit for him)

Credit-to: Erkka Kääriä <erkka.kaaria@intel.com>